requirements.txt
groq>=0.9.0
psycopg2-binary>=2.9.9
pymysql>=1.1.0
sqlparse>=0.5.0
rich>=13.7.0
python-dotenv>=1.0.0
